extracted fade in monitor preset changers to pure component
