+++ src/screens/TestMonitorScreen.js
@@ -6,6 +6,7 @@ import Orientation from "react-native-orientation";
 import { NetworkComp } from '../components/network';
 import PhilipsMonitor from '../components/MonitorPresets/PhilipsMonitor';
 import FadeInView from 'react-native-fade-in-view';
+import PresetChangerArrow from '../components/MonitorPresets/PresetChangerArrow';
 
 class MonitorScreen extends Component {
     constructor(props) {
@@ -39,37 +40,20 @@ class MonitorScreen extends Component {
                 style={{flex: 1, flexDirection: 'row'}}
                 onResponderRelease={() => this.stopTouch(this.state.toggle)}
                 onStartShouldSetResponder={(e) => {return true}}>
-                {this.state.toggle &&
-                <FadeInView 
-                    duration={750}
-                    style={[styles.presetChangers,
-                    {left: Dimensions.get('window').width*(1/15),
-                     top: Dimensions.get('window').height*(1/2.5)}]}>
-                    <TouchableOpacity style={styles.presetChangerButton}>
-                        <Icon
-                            type='MaterialIcons'
-                            name='chevron-left'
-                            style={{color: 'white'}}
-                        />
-                    </TouchableOpacity>
-                </FadeInView>
-                }
-                {this.state.toggle &&
-                <FadeInView
-                    duration={750}
-                    style={[styles.presetChangers,
-                        {left: Dimensions.get('window').width*(14/15),
-                         top: Dimensions.get('window').height*(1/2.5)-5}]}>
-                    <TouchableOpacity style={styles.presetChangerButton}>
-                        <Icon
-                        type='MaterialIcons'
-                        name='chevron-right'
-                        style={{color: 'white'}} />
-                    </TouchableOpacity>
-                </FadeInView>
-                }
                 <NetworkComp />
                 <PhilipsMonitor />
+                <PresetChangerArrow
+                    show={this.state.toggle}
+                    arrow="chevron-left"
+                    style={{left: Dimensions.get('window').width*(1/15),
+                            top: Dimensions.get('window').height*(1/2.5)}}
+                />
+                <PresetChangerArrow
+                    show={this.state.toggle}
+                    arrow="chevron-right"
+                    style={{left: Dimensions.get('window').width*(14/15),
+                            top: Dimensions.get('window').height*(1/2.5)-5}}
+                />
             </View>
         );
     }
@@ -84,20 +68,3 @@ const mapStateToProps = (state) => {
 }
 
 export default connect(mapStateToProps)(MonitorScreen);
-
-const styles = StyleSheet.create({
-    presetChangers: {
-        flex: 1,
-        position: 'absolute',
-        width: 25,
-        height: 80,
-        zIndex: 100
-    },
-    presetChangerButton: {
-        flex: 1,
-        backgroundColor: 'white',
-        opacity: 0.4,
-        justifyContent: 'center',
-        borderRadius: 5
-    }
-});
